Fixed two exceptions that were thrown.
1. Replaced the reference to class Random with the full classpath java.util.Random,
and removed '(java.util Random) from the argument of import on line 3. An exception
was thrown because a class called Random was already in the current namespace.

2. Modified the manner in which integers are converted into bytes by introducing
a function int-to-ubyte. This fixes an exception triggered when evaluating (byte i)
in the body of dotimes within the definition of make-channel in make-grayscale-lut.